Doc-BOT/requirements.txt
Flask
Flask-Cors
edge-tts
google-generativeai
SpeechRecognition
pydub